+++ src/main/java/tg/DocVers/Repository/DocInfoRepository.java
@@ -1,9 +1,10 @@
 package tg.DocVers.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import tg.DocVers.Entity.DocInfo;
 
 @Repository
-public interface DocInfoRepository extends JpaRepository<DocInfo, Integer> {
+public interface DocInfoRepository extends JpaRepository<DocInfo, Long> {
 }
